Added comments, made paths more OS agnostic
